+++ voicevox_engine/tts_pipeline/tts_engine.py
@@ -9,7 +9,7 @@ from soxr import resample
 from .acoustic_feature_extractor import OjtPhoneme
 from ..model import AccentPhrase, AudioQuery, Mora
 from .core_wrapper import CoreWrapper, OldCoreError
-from .synthesis_engine_base import SynthesisEngineBase
+from .tts_engine_base import SynthesisEngineBase
 
 unvoiced_mora_phoneme_list = ["A", "I", "U", "E", "O", "cl", "pau"]
 mora_phoneme_list = ["a", "i", "u", "e", "o", "N"] + unvoiced_mora_phoneme_list
